requirements.txt
amqp==5.2.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
celery==5.4.0
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
Django==5.0.6
django-celery-beat==2.6.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3
django-jazzmin==3.0.0
django-redis==5.4.0
django-timezone-field==6.1.0
djangorestframework==3.15.1
drf-yasg==1.21.7
exceptiongroup==1.2.1
factory-boy==3.3.0
Faker==25.3.0
flower==2.0.1
humanize==4.9.0
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
kombu==5.3.7
MarkupSafe==2.1.5
packaging==24.0
pluggy==1.5.0
prometheus_client==0.20.0
prompt_toolkit==3.0.45
psycopg==3.1.19
psycopg-c==3.1.19
pycparser==2.22
pytest==8.2.1
pytest-django==4.8.0
pytest-sugar==1.0.0
python-crontab==3.1.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
termcolor==2.4.0
text-unidecode==1.3
tomli==2.0.1
tornado==6.4
typing_extensions==4.12.1
tzdata==2024.1
uritemplate==4.1.1
vine==5.1.0
watchdog==4.0.1
watchfiles==0.22.0
wcwidth==0.2.13
Werkzeug==3.0.3
